requirements.txt
fastapi==0.85.1
fire==0.4.0
gunicorn==20.1.0
langdetect==1.0.9
nltk==3.7
pip-chill==1.0.1
phonenumbers==8.13.0
python-dotenv==0.21.0
python-levenshtein==0.20.8
textblob==0.17.1
textstat==0.7.3
translate==3.6.1
uvicorn==0.19.0
scikit-learn==1.1.3
spacy==3.4.2
quantulum3==0.7.11
language_tool_python==2.7.1
LeXmo==0.1.4
better_profanity==0.7.0